services/agentura-cron/requirements.txt
apscheduler>=3.10.0
httpx>=0.27.0
pyyaml>=6.0
